+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2026-01-31
+### Added
+- Add New SetForegroundColor Actions
+
 ## 2026-01-23
 ### Changed
 - Screen Color Picker is now only loaded on Windows OS